pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.30.0",
    "matplotlib>=3.10.5",
    "pandasql>=0.7.3",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
